Organize Attention Backends (#1547)
